test(stripe): split checkout scenarios into checkout.test.ts
Pulls 6 sub-describes (line item replacement / asymmetric / duplicate
prevention, subscriptionSuccess - checkoutSessionId flow, metered
usage pricing, getCheckoutSessionParams subscription_data merge) out
of stripe.test.ts. Wraps them in a "stripe checkout" describe.
stripe.test.ts loses ~2,210 lines.
